Allow using external propensities in estimate_ate when pretrain=True without triggering a retrain of the model
When pretrain=True and p is not None, the implementation of estimate_ate triggered a retrain of the model. I believe this is not the expected behavior, so I changed it. When the Xlearner is used with externally generated propensity scores, one should be able to use estimate_ate with the same externally generated propensity scores without retaining the model. Now when pretrain=True and p is not None, estimate_ate will use the pretrained model and the vector of propensities p.